refactor: move timingSafeEqual out of worker-utils to encryption follow-up PR
Restore local timingSafeEqual implementations in cloudflare-db-proxy,
cloudflare-webhook-agent-ingest, and kiloclaw to keep this PR focused
on non-crypto utilities. The shared extraction will land in PR #697.
