+++ kiloclaw/src/auth/middleware.ts
@@ -1,6 +1,5 @@
 import type { Context, Next } from 'hono';
 import { getCookie } from 'hono/cookie';
-import { timingSafeEqual } from '@kilocode/worker-utils';
 import type { AppEnv } from '../types';
 import { KILOCLAW_AUTH_COOKIE } from '../config';
 import { validateKiloToken } from './jwt';
@@ -101,4 +100,17 @@ export async function internalApiMiddleware(c: Context<AppEnv>, next: Next) {
   }
 
   return next();
+}
+
+function timingSafeEqual(a: string, b: string): boolean {
+  const encoder = new TextEncoder();
+  const aBytes = encoder.encode(a);
+  const bBytes = encoder.encode(b);
+
+  if (aBytes.length !== bBytes.length) {
+    // Compare a against itself so the timing is constant regardless of length mismatch
+    crypto.subtle.timingSafeEqual(aBytes, aBytes);
+    return false;
+  }
+  return crypto.subtle.timingSafeEqual(aBytes, bBytes);
 }
