+++ packages/worker-utils/src/index.ts
@@ -1,38 +1,24 @@
-// 1. DO retry
 export { withDORetry, DEFAULT_DO_RETRY_CONFIG } from './do-retry.js';
 export type { DORetryConfig } from './do-retry.js';
 
-// 2. Backend auth middleware
 export { backendAuthMiddleware } from './backend-auth-middleware.js';
 
-// 3. Timeout
 export { withTimeout } from './timeout.js';
 
-// 4. R2 client
 export { createR2Client } from './r2-client.js';
 export type { R2Client, R2ClientConfig } from './r2-client.js';
 
-// 5. Response helpers
 export { resSuccess, resError } from './res.js';
 export type { SuccessResponse, ErrorResponse, ApiResponse } from './res.js';
 
-// 6. Zod JSON validator
 export { zodJsonValidator } from './zod-json-validator.js';
 
-// 7. Timing-safe equal
-export { timingSafeEqual } from './timing-safe-equal.js';
-
-// 8. Format error
 export { formatError } from './format-error.js';
 
-// 9. Extract bearer token
 export { extractBearerToken } from './extract-bearer-token.js';
 
-// 10. Error handler
 export { createErrorHandler } from './error-handler.js';
 
-// 11. Not-found handler
 export { createNotFoundHandler } from './not-found-handler.js';
 
-// 12-13. Shared types
 export type { Owner, MCPServerConfig } from './types.js';
